+++ console/algorithms/src/poseidon/helpers/mod.rs
@@ -40,7 +40,7 @@ pub trait AlgebraicSponge<F: PrimeField, const RATE: usize, const CAPACITY: usiz
     fn absorb(&mut self, input: &[F]);
 
     /// Squeeze `num_elements` field elements from the sponge.
-    fn squeeze(&mut self, num_elements: usize) -> SmallVec<[F; 10]>;
+    fn squeeze(&mut self, num_elements: u16) -> SmallVec<[F; 10]>;
 }
 
 pub trait DefaultCapacityAlgebraicSponge<F: PrimeField, const RATE: usize>: AlgebraicSponge<F, RATE, 1> {
